fix test

+++ crates/serverless/src/deployments/filesystem.rs
@@ -16,8 +16,8 @@ pub fn create_deployments_folder() -> Result<()> {
 }
 
 pub fn rm_deployment(deployment_id: &str) -> Result<()> {
-    #[cfg(not(test))]
-    {
+    let is_test = std::env::var("LAGOSS_TEST").is_ok() || cfg!(test);
+    if !is_test {
         let path = Path::new(DEPLOYMENTS_DIR).join(deployment_id);
         if path.exists() {
             fs::remove_dir_all(path)?;
